+++ client/client.cabal
@@ -90,10 +90,16 @@ library
     , transformers
     , hsqml
     , text
+    , network
+    , binary
+    , bytestring
     , uuid-types
   exposed-modules:
       CLenses
       ClientTypes
+      GUISetup
+      HandleMessage
+      Utilities
   other-modules:
       Paths_client
   default-language: Haskell2010
@@ -109,8 +115,8 @@ executable client
     , transformers
     , hsqml
     , text
-    , client
     , network
-    , bytestring
     , binary
+    , bytestring
+    , client
   default-language: Haskell2010
