+++ client/package.yaml
@@ -15,6 +15,9 @@ dependencies:
 - transformers
 - hsqml
 - text
+- network
+- binary
+- bytestring
 library:
     source-dirs: src/lib
     dependencies:
@@ -25,6 +28,3 @@ executables:
         source-dirs: src/exe
         dependencies:
         - client
-        - network
-        - bytestring
-        - binary
